iter57 cluster-072 fix r2: 删 production-unused serializer + guard 切真路径(actor-backed catalog stores)
PR #956 r1 reject 2 blockers:
- ConnectorCatalogStorageSerializer + RoleCatalogStorageSerializer 不在 production path(只 tests + guard)
- guard 只扫这 2 unused files,真路径 ActorBackedXxxStore + import parsers 没覆盖

修(plan B 删壳+真 guard):
- 删 unused serializer + 相应 tests
- guard 改覆盖 ActorBackedConnectorCatalogStore / ActorBackedRoleCatalogStore + StudioLocalCatalogImportReader + import parsers
- 验:加临时 JSON 入侵真路径 → guard fail;移除 → pass

⟦AI:AUTO-LOOP⟧
